refactor: migrate calculator buttons and header buttons to separate UserControls
