+++ CyberCalc.NET9/Controls/ButtonsControl.xaml
@@ -1,4 +1,4 @@
-<UserControl x:Class="WPF_CALC_NET_9.Views.ButtonsControl"
+<UserControl x:Class="WPF_CALC_NET_9.Controls.ButtonsControl"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
