+++ CyberCalc.NET9/Controls/ButtonsControl.xaml.cs
@@ -1,6 +1,7 @@
+// ButtonsControl.xaml.cs
 using System.Windows.Controls;
 
-namespace WPF_CALC_NET_9.Views
+namespace WPF_CALC_NET_9.Controls
 {
     public partial class ButtonsControl : UserControl
     {
